added functions to break down level array
now with an array of beats the game can produce a correct number of
tracks. and the OO structure for level and tracks works beautifully.
Also added functionality where when a platform lights up its tracks
corresponding sound is played. Not synced to bouncing of ball yet but
getting there.

Interesting note: if you check the level file i hardcopied the level1
array. Because i had too. spent way too much time trying to get
processing to use the array from javascript. It would be really cool if
we could figure this out. PLEASE!!!!